requirements.txt
aiohttp
websockets
pyyaml
python-dotenv
aiosqlite
asyncpg
pytest
pytest-asyncio
aioresponses

